app1/register-face/requirements.txt
torch
torchvision
opencv-python-headless
mysql-connector-python
numpy
requests

